ai-engine/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
redis==4.6.0
python-dotenv==1.0.0
httpx==0.25.0
asyncio==3.4.3
aioredis==2.0.1
pytest==7.4.2
psycopg2-binary==2.9.7
sqlalchemy==2.0.21
alembic==1.12.0
python-multipart==0.0.6
jinja2==3.1.2
plotly==5.17.0
seaborn==0.12.2
matplotlib==3.7.2
xgboost==1.7.6
lightgbm==4.1.0
joblib==1.3.2
scipy==1.11.3